Factor out an email mixin
